GET, PUT, POST, DELETE Fully Added

+++ src/main/java/com/sraynitjsr/Student/Management/System/repository/StudentRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
+
 }
